fix: move chart/constant into root dir, keep constant file single
