+++ src/chart/controller/legend.ts
@@ -1,7 +1,7 @@
 import * as _ from '@antv/util';
+import { ComponentType, DIRECTION, LAYER } from '../../constant';
 import { Attribute, Scale } from '../../dependents';
 import { Legend } from '../__components__';
-import { ComponentType, DIRECTION, LAYER } from '../constant';
 import { ComponentOption, LegendCfg } from '../interface';
 import View from '../view';
