+++ src/chart/index.ts
@@ -1,5 +1,4 @@
 export { default as Chart } from './chart';
 export { default as View, registerGeometry } from './view';
 
-export { LAYER, DIRECTION } from './constant';
 export { ChartCfg, ViewCfg } from './interface';
