+++ src/interface.ts
@@ -1,3 +1,10 @@
+/** G 的渲染类型 */
+export type Renderer = 'svg' | 'canvas';
+
+/** 数据的定义 */
+export type Datum = Record<string, string | number>;
+export type Data = Datum[];
+
 /** 对象 */
 export interface LooseObject {
   [key: string]: any;
