+++ tests/unit/util/direction-spec.ts
@@ -1,5 +1,5 @@
 import BBox from '@antv/g/lib/core/bbox';
-import { DIRECTION } from '../../../src/chart';
+import { DIRECTION } from '../../../src';
 import { directionToPosition } from '../../../src/util/direction';
 
 const parentBBox = new BBox(10, 10, 100, 100);
